+++ Tests/DetermineArtifactUrl.Test.ps1
@@ -72,7 +72,6 @@ Describe "DetermineArtifactUrl" {
             'country' = 'us'
             'applicationDependency' = '20.0.0.0'
             'additionalCountries' = @()
-            'useEarliestArtifact' = $false
         }
     }
 
@@ -114,10 +113,16 @@ Describe "DetermineArtifactUrl" {
         DetermineArtifactUrl -projectSettings $projectSettings | should -be 'https://bcartifacts/sandbox/22.1.12345.12345/us'
     }
 
-    It 'Artifact setting when useEarliestArtifact' {
+    It 'Artifact setting when using version = * and first' {
         $projectSettings.applicationDependency = '22.1.0.0'
-        $projectSettings.useEarliestArtifact = $true
+        $projectSettings.artifact = "//*/us/first"
         DetermineArtifactUrl -projectSettings $projectSettings | should -be 'https://bcartifacts/sandbox/22.1.12345.12345/us'
+    }
+
+    It 'Artifact setting when using version = * and latest' {
+        $projectSettings.applicationDependency = '22.1.0.0'
+        $projectSettings.artifact = "//*/us/latest"
+        DetermineArtifactUrl -projectSettings $projectSettings | should -be 'https://bcartifacts/sandbox/22.1.12345.12346/us'
     }
 }